+++ cogs/Roles.py
@@ -3,8 +3,8 @@ import json
 import discord
 from discord.ext import commands
 
-with open("./etc/config.json", "r") as f:
-    config = json.load(f)
+with open("./etc/config.json", "r") as f_org:
+    config = json.load(f_org)
 
 
 class roles(commands.Cog):
@@ -23,23 +23,32 @@ class roles(commands.Cog):
         - **?gawrole**
         """
 
-        role = discord.utils.get(ctx.guild.roles, name='Giveaway')
-        if not role:
-            try:
-                giveawayrole = await ctx.guild.create_role(name='Giveaway', color=discord.Color.green())
-                embed = discord.Embed(title='<:open:869959941321011260> Successfully',
-                                      description=f'Die Rolle **{giveawayrole}** wurde erstellt!')
+        with open('utils/json/active_check.json', 'r') as f:
+            data = json.load(f)
+
+        if data[str(ctx.guild.id)]["Roles"] == 'false':
+            embed = discord.Embed(
+                description=f'Diese **Extension (Roles) ist momentan deaktiviert!** Wende dich bitte an **den Owner vom Bot** (LookAtYourSkill#6666)',
+                color=discord.Color.red())
+            await ctx.send(embed=embed)
+        else:
+            role = discord.utils.get(ctx.guild.roles, name='Giveaway')
+            if not role:
+                try:
+                    giveawayrole = await ctx.guild.create_role(name='Giveaway', color=discord.Color.green())
+                    embed = discord.Embed(title='<:open:869959941321011260> Successfully',
+                                          description=f'Die Rolle **{giveawayrole}** wurde erstellt!')
+                    await ctx.send(embed=embed, delete_after=5)
+                    await asyncio.sleep(1)
+                    await ctx.message.delete()
+                except discord.Forbidden:
+                    pass
+            else:
+                embed = discord.Embed(title='<:close:864599591692009513> **ERROR**',
+                                      description=f'Die Rolle **existiert bereits**!')
                 await ctx.send(embed=embed, delete_after=5)
                 await asyncio.sleep(1)
                 await ctx.message.delete()
-            except discord.Forbidden:
-                pass
-        else:
-            embed = discord.Embed(title='<:close:864599591692009513> **ERROR**',
-                                  description=f'Die Rolle **existiert bereits**!')
-            await ctx.send(embed=embed, delete_after=5)
-            await asyncio.sleep(1)
-            await ctx.message.delete()
 
     @commands.command(name='muterole', aliases=['createMute'])
     @commands.has_permissions(manage_roles=True)
@@ -49,30 +58,39 @@ class roles(commands.Cog):
         - **?muterole**
         """
 
-        role = discord.utils.get(ctx.guild.roles, name='Muted')
-        if not role:
-            try:
-                muterole = await ctx.guild.create_role(name='Muted', color=discord.Color.darker_gray())
-                embed = discord.Embed(title='<:open:869959941321011260> Successful',
-                                      description=f'Die Rolle **{muterole}** wurde erstellt!')
+        with open('utils/json/active_check.json', 'r') as f:
+            data = json.load(f)
+
+        if data[str(ctx.guild.id)]["Roles"] == 'false':
+            embed = discord.Embed(
+                description=f'Diese **Extension (Roles) ist momentan deaktiviert!** Wende dich bitte an **den Owner vom Bot** (LookAtYourSkill#6666)',
+                color=discord.Color.red())
+            await ctx.send(embed=embed)
+        else:
+            role = discord.utils.get(ctx.guild.roles, name='Muted')
+            if not role:
+                try:
+                    muterole = await ctx.guild.create_role(name='Muted', color=discord.Color.darker_gray())
+                    embed = discord.Embed(title='<:open:869959941321011260> Successful',
+                                          description=f'Die Rolle **{muterole}** wurde erstellt!')
+                    await ctx.send(embed=embed, delete_after=5)
+                    await asyncio.sleep(1)
+                    await ctx.message.delete()
+                    for channel in ctx.guild.channels:
+                        await channel.set_permissions(muterole,
+                                                      speak=False,
+                                                      send_messages=False,
+                                                      read_messages=True,
+                                                      read_message_history=True,
+                                                      )
+                except discord.Forbidden:
+                    pass
+            else:
+                embed = discord.Embed(title='<:close:864599591692009513> **ERROR**',
+                                      description=f'Die Rolle **existiert bereits**!')
                 await ctx.send(embed=embed, delete_after=5)
                 await asyncio.sleep(1)
                 await ctx.message.delete()
-                for channel in ctx.guild.channels:
-                    await channel.set_permissions(muterole,
-                                                  speak=False,
-                                                  send_messages=False,
-                                                  read_messages=True,
-                                                  read_message_history=True,
-                                                  )
-            except discord.Forbidden:
-                pass
-        else:
-            embed = discord.Embed(title='<:close:864599591692009513> **ERROR**',
-                                  description=f'Die Rolle **existiert bereits**!')
-            await ctx.send(embed=embed, delete_after=5)
-            await asyncio.sleep(1)
-            await ctx.message.delete()
 
     @commands.command(name='giverole')
     @commands.has_permissions(manage_roles=True)
@@ -82,15 +100,23 @@ class roles(commands.Cog):
         - **?giverole [`user`] [`roleid`]**
         """
 
-        try:
-            await user.add_roles(role)
-            embed = discord.Embed(title='',
-                                  description=f'Dem User **{user}** wurde die Rolle `{role} gegeben!`')
-            await ctx.send(embed=embed, delete_after=5)
-            await asyncio.sleep(1)
-            await ctx.message.delete()
-        except discord.Forbidden:
-            pass
+        with open('utils/json/active_check.json', 'r') as f:
+            data = json.load(f)
+
+        if data[str(ctx.guild.id)]["Roles"] == 'false':
+            embed = discord.Embed(
+                description=f'Diese **Extension (Roles) ist momentan deaktiviert!** Wende dich bitte an **den Owner vom Bot** (LookAtYourSkill#6666)',
+                color=discord.Color.red())
+            await ctx.send(embed=embed)
+        else:
+            try:
+                await user.add_roles(role)
+                embed = discord.Embed(description=f'Dem User **{user}** wurde die Rolle `{role} gegeben!`')
+                await ctx.send(embed=embed, delete_after=5)
+                await asyncio.sleep(1)
+                await ctx.message.delete()
+            except discord.Forbidden:
+                pass
 
     @commands.command(name='removerole', aliases=['rmrole'])
     @commands.has_permissions(manage_roles=True)
@@ -100,15 +126,23 @@ class roles(commands.Cog):
         - **?rmrole [`user`] [`roleid`]**
         """
 
-        try:
-            await user.remove_roles(role)
-            embed = discord.Embed(title='',
-                                  description=f'Dem User **{user}** wurde die Rolle `{role} entfernt!`')
-            await ctx.send(embed=embed, delete_after=5)
-            await asyncio.sleep(1)
-            await ctx.message.delete()
-        except discord.Forbidden:
-            pass
+        with open('utils/json/active_check.json', 'r') as f:
+            data = json.load(f)
+
+        if data[str(ctx.guild.id)]["Roles"] == 'false':
+            embed = discord.Embed(
+                description=f'Diese **Extension (Roles) ist momentan deaktiviert!** Wende dich bitte an **den Owner vom Bot** (LookAtYourSkill#6666)',
+                color=discord.Color.red())
+            await ctx.send(embed=embed)
+        else:
+            try:
+                await user.remove_roles(role)
+                embed = discord.Embed(description=f'Dem User **{user}** wurde die Rolle `{role} entfernt!`')
+                await ctx.send(embed=embed, delete_after=5)
+                await asyncio.sleep(1)
+                await ctx.message.delete()
+            except discord.Forbidden:
+                pass
 
     @commands.command(aliases=['createrole'])
     @commands.has_permissions(manage_roles=True)
@@ -118,12 +152,21 @@ class roles(commands.Cog):
         - **?createrole [`role_name`]**
         """
 
-        await ctx.guild.create_role(name=role_name, color=color)
-        embed = discord.Embed(title='<:open:869959941321011260> Successfully',
-                              description=f'Die Rolle **{role_name}** wurde erstellt!')
-        await ctx.send(embed=embed, delete_after=5)
-        await asyncio.sleep(1)
-        await ctx.message.delete()
+        with open('utils/json/active_check.json', 'r') as f:
+            data = json.load(f)
+
+        if data[str(ctx.guild.id)]["Roles"] == 'false':
+            embed = discord.Embed(
+                description=f'Diese **Extension (Roles) ist momentan deaktiviert!** Wende dich bitte an **den Owner vom Bot** (LookAtYourSkill#6666)',
+                color=discord.Color.red())
+            await ctx.send(embed=embed)
+        else:
+            await ctx.guild.create_role(name=role_name, color=color)
+            embed = discord.Embed(title='<:open:869959941321011260> Successfully',
+                                  description=f'Die Rolle **{role_name}** wurde erstellt!')
+            await ctx.send(embed=embed, delete_after=5)
+            await asyncio.sleep(1)
+            await ctx.message.delete()
 
     @commands.command(aliases=['delrole'])
     @commands.has_permissions(manage_roles=True)
@@ -133,23 +176,32 @@ class roles(commands.Cog):
         - **?delrole [`role_name`]**
         """
 
-        role = discord.utils.get(ctx.guild.roles, name=role_name)
-        if role:
-            try:
-                await role.delete()
-                embed = discord.Embed(title='<:open:869959941321011260> Successfully',
-                                      description=f'Die Rolle **{role_name}** wurde gelöscht!')
+        with open('utils/json/active_check.json', 'r') as f:
+            data = json.load(f)
+
+        if data[str(ctx.guild.id)]["Roles"] == 'false':
+            embed = discord.Embed(
+                description=f'Diese **Extension (Roles) ist momentan deaktiviert!** Wende dich bitte an **den Owner vom Bot** (LookAtYourSkill#6666)',
+                color=discord.Color.red())
+            await ctx.send(embed=embed)
+        else:
+            role = discord.utils.get(ctx.guild.roles, name=role_name)
+            if role:
+                try:
+                    await role.delete()
+                    embed = discord.Embed(title='<:open:869959941321011260> Successfully',
+                                          description=f'Die Rolle **{role_name}** wurde gelöscht!')
+                    await ctx.send(embed=embed, delete_after=5)
+                    await asyncio.sleep(1)
+                    await ctx.message.delete()
+                except discord.Forbidden:
+                    pass
+            else:
+                embed = discord.Embed(title='<:close:864599591692009513> **ERROR**',
+                                      description=f'Die Rolle **existiert nicht**!')
                 await ctx.send(embed=embed, delete_after=5)
                 await asyncio.sleep(1)
                 await ctx.message.delete()
-            except discord.Forbidden:
-                pass
-        else:
-            embed = discord.Embed(title='<:close:864599591692009513> **ERROR**',
-                                  description=f'Die Rolle **existiert nicht**!')
-            await ctx.send(embed=embed, delete_after=5)
-            await asyncio.sleep(1)
-            await ctx.message.delete()
 
     @commands.command(name='getroles', aliases=['getr'])
     @commands.has_permissions(manage_roles=True)
@@ -159,18 +211,27 @@ class roles(commands.Cog):
         - **?getr [`member`]**
         """
 
-        if not member:
-            member = ctx.author
+        with open('utils/json/active_check.json', 'r') as f:
+            data = json.load(f)
 
-        embed = discord.Embed(title=f'Roles from {member}')
-        role_list = ''
-        for i in member.roles:
-            role_list += f'- `{i}`\n'
+        if data[str(ctx.guild.id)]["Roles"] == 'false':
+            embed = discord.Embed(
+                description=f'Diese **Extension (Roles) ist momentan deaktiviert!** Wende dich bitte an **den Owner vom Bot** (LookAtYourSkill#6666)',
+                color=discord.Color.red())
+            await ctx.send(embed=embed)
+        else:
+            if not member:
+                member = ctx.author
 
-        embed.add_field(name='Getted Role',
-                        value=f'{role_list}',
-                        inline=False)
-        await ctx.send(embed=embed)
+            embed = discord.Embed(title=f'Roles from {member}')
+            role_list = ''
+            for i in member.roles:
+                role_list += f'- `{i}`\n'
+
+            embed.add_field(name='Getted Role',
+                            value=f'{role_list}',
+                            inline=False)
+            await ctx.send(embed=embed)
 
     @commands.command(name='removeroles', aliases=['remover'])
     @commands.has_permissions(manage_roles=True)
@@ -180,25 +241,34 @@ class roles(commands.Cog):
         - **?removeroles [`member`]**
         """
 
-        if not member:
-            embed = discord.Embed(description=f'Du kannst dir nicht selbst alle Rollen wegnehmen...')
+        with open('utils/json/active_check.json', 'r') as f:
+            data = json.load(f)
+
+        if data[str(ctx.guild.id)]["Roles"] == 'false':
+            embed = discord.Embed(
+                description=f'Diese **Extension (Roles) ist momentan deaktiviert!** Wende dich bitte an **den Owner vom Bot** (LookAtYourSkill#6666)',
+                color=discord.Color.red())
             await ctx.send(embed=embed)
-            return
+        else:
+            if not member:
+                embed = discord.Embed(description=f'Du kannst dir nicht selbst alle Rollen wegnehmen...')
+                await ctx.send(embed=embed)
+                return
 
-        members_roles = member.roles
+            members_roles = member.roles
 
-        for i in range(len(members_roles) - 1):
-            await member.remove_roles(members_roles[i + 1])
+            for i in range(len(members_roles) - 1):
+                await member.remove_roles(members_roles[i + 1])
 
-            if len(members_roles) == 1:
-                await ctx.send(f'The member has no roles. {member.mention} should do the ?verify command!')
+                if len(members_roles) == 1:
+                    await ctx.send(f'The member has no roles. {member.mention} should do the ?verify command!')
 
-            else:
-                role = discord.utils.get(ctx.guild.roles, id=916860116207280159)
-                await member.add_roles(role)
-                await member.remove_roles(role)
+                else:
+                    role = discord.utils.get(ctx.guild.roles, id=916860116207280159)
+                    await member.add_roles(role)
+                    await member.remove_roles(role)
 
-        await ctx.send(f'Removed {len(members_roles) - 1} from {member.mention} successful!')
+            await ctx.send(f'Removed {len(members_roles) - 1} from {member.mention} successful!')
 
 
 def setup(bot):
